refactor: chat 도메인 presentation 계층 정리
