+++ src/main/java/umc/cockple/demo/domain/chat/presentation/websocket/ChatWebSocketConfig.java
@@ -1,17 +1,15 @@
-package umc.cockple.demo.global.config;
+package umc.cockple.demo.domain.chat.presentation.websocket;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.socket.config.annotation.EnableWebSocket;
 import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
 import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
-import umc.cockple.demo.domain.chat.handler.ChatWebSocketHandler;
-import umc.cockple.demo.domain.chat.interceptor.JWTWebSocketAuthInterceptor;
 
 @Configuration
 @EnableWebSocket
 @RequiredArgsConstructor
-public class WebSocketConfig implements WebSocketConfigurer {
+public class ChatWebSocketConfig implements WebSocketConfigurer {
 
     private final ChatWebSocketHandler chatWebSocketHandler;
     private final JWTWebSocketAuthInterceptor jwtWebSocketAuthInterceptor;
