+++ src/main/java/umc/cockple/demo/domain/chat/presentation/websocket/JWTWebSocketAuthInterceptor.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.domain.chat.interceptor;
+package umc.cockple.demo.domain.chat.presentation.websocket;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
